+++ pytket/extensions/qiskit/qiskit_convert.py
@@ -19,7 +19,6 @@ from collections import defaultdict
 from typing import (
     Callable,
     Optional,
-    Union,
     Any,
     Iterable,
     cast,
@@ -61,7 +60,8 @@ from qiskit.circuit.library import (
     UnitaryGate,
     Initialize,
 )
-from qiskit.providers.models import BackendProperties, QasmBackendConfiguration  # type: ignore
+from qiskit_ibm_runtime.models.backend_configuration import PulseBackendConfiguration  # type: ignore
+from qiskit_ibm_runtime.models.backend_properties import BackendProperties  # type: ignore
 
 from pytket.circuit import (
     CircBox,
@@ -82,11 +82,11 @@ from pytket.unit_id import _TEMP_BIT_NAME
 from pytket.pauli import Pauli, QubitPauliString
 from pytket.architecture import Architecture, FullyConnected
 from pytket.utils import QubitPauliOperator, gen_term_sequence_circuit
-from pytket.passes import RebaseCustom
+from pytket.passes import AutoRebase
 
 if TYPE_CHECKING:
-    from qiskit.providers.backend import BackendV1  # type: ignore
-    from qiskit.providers.models.backendproperties import Nduv  # type: ignore
+    from qiskit_ibm_runtime.ibm_backend import IBMBackend  # type: ignore
+    from qiskit_ibm_runtime.models.backend_properties import Nduv
     from qiskit.circuit.quantumcircuitdata import QuantumCircuitData  # type: ignore
     from pytket.circuit import Op, UnitID
 
@@ -205,7 +205,7 @@ _gate_str_2_optype_rev = {v: k for k, v in _gate_str_2_optype.items()}
 _gate_str_2_optype_rev[OpType.Unitary1qBox] = "unitary"
 
 
-def _tk_gate_set(config: QasmBackendConfiguration) -> set[OpType]:
+def _tk_gate_set(config: PulseBackendConfiguration) -> set[OpType]:
     """Set of tket gate types supported by the qiskit backend"""
     if config.simulator:
         gate_set = {
@@ -805,9 +805,6 @@ order or only one bit of one register"""
     return qcirc.append(g, qargs=qargs)
 
 
-# Define varibles for RebaseCustom
-_cx_replacement = Circuit(2).CX(0, 1)
-
 # The set of tket gates that can be converted directly to qiskit gates
 _supported_tket_gates = set(_known_gate_rev_phase.keys())
 
@@ -821,19 +818,8 @@ _protected_tket_gates = (
     | {OpType.CustomGate}
 )
 
-
-Param = Union[float, "sympy.Expr"]  # Type for TK1 and U3 parameters
-
-
-# Use the U3 gate for tk1_replacement as this is a member of _supported_tket_gates
-def _tk1_to_u3(a: Param, b: Param, c: Param) -> Circuit:
-    tk1_circ = Circuit(1)
-    tk1_circ.U3(b, a - 1 / 2, c + 1 / 2, 0).add_phase(-(a + c) / 2)
-    return tk1_circ
-
-
 # This is a rebase to the set of tket gates which have an exact substitution in qiskit
-supported_gate_rebase = RebaseCustom(_protected_tket_gates, _cx_replacement, _tk1_to_u3)
+supported_gate_rebase = AutoRebase(_protected_tket_gates)
 
 
 def tk_to_qiskit(
@@ -907,9 +893,9 @@ def tk_to_qiskit(
     return qcirc
 
 
-def process_characterisation(backend: "BackendV1") -> dict[str, Any]:
-    """Convert a :py:class:`qiskit.providers.backend.BackendV1` to a dictionary
-     containing device Characteristics
+def process_characterisation(backend: "IBMBackend") -> dict[str, Any]:
+    """Convert a :py:class:`qiskit_ibm_runtime.ibm_backend.IBMBackend` to a
+    dictionary containing device Characteristics
 
     :param backend: A backend to be converted
     :return: A dictionary containing device characteristics
@@ -920,7 +906,7 @@ def process_characterisation(backend: "BackendV1") -> dict[str, Any]:
 
 
 def process_characterisation_from_config(
-    config: QasmBackendConfiguration, properties: Optional[BackendProperties]
+    config: PulseBackendConfiguration, properties: Optional[BackendProperties]
 ) -> dict[str, Any]:
     """Obtain a dictionary containing device Characteristics given config and props.